支持二代数据全聚焦数据，支持tofd真实深度，支持两台仪器远程监控

+++ phascan_view/configure/config_phascan_ii/mercury_paramters/wedge_paramters.h
@@ -39,7 +39,7 @@ private:
     double m_clampOffset;
 
     WaveType m_waveType;
-    uint m_refPoint;
+    double m_refPoint;
 
     friend class Config;
 };
